Fix snake growth bug in UI
Old snake parts were not being shifted properly in the UI despite being
popped properly internally in the handleMove. Cause was the screen fill
being outside of the updateUI, so when the snake moved nothing was in
place to "paint over" the old snake elements that were a part of its
past body.